requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
tmdbv3api==1.9.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0